Big Commit
Changes such as:
Edits
Metadata
etc.

+++ MMS-Portal/Database/Forms/DeleteCause/server.php
@@ -8,15 +8,56 @@
  */
 include_once './Database/DAO/CauseDB.php';
 include_once './Database/DAO/EffectDB.php';
+include_once './Database/DAO/ClusterDB.php';
+include_once './Database/DAO/CauseEffectDB.php';
 
 //Delete user
+function errorHandlingDeleteEffect(){
 if (isset($_POST['delete_effect']) && isset($_POST['delete_idEffect'])) {
-    EffectDB::deleteById($_POST['delete_idEffect']);
+    
+        if(ClusterDB::getSearchClusterEffects($_POST['delete_idEffect']) && CauseEffectDB::getCausebyEffectIdOne($_POST['delete_idEffect'])){
+            
+            echo '<span style="color:red">Delete denied: Effect exists in Cluster and in Cause-Effect</span>';
+            
+        }else if(CauseEffectDB::getCausebyEffectIdOne($_POST['delete_idEffect'])){
+            
+            echo '<span style="color:red">Delete denied: Effect exists in Cause-Effect</span>';
+            
+        } else if(ClusterDB::getSearchClusterEffects($_POST['delete_idEffect'])) {
+            
+            echo '<span style="color:red">Delete denied: Effect exists in Cluster</span>';
+    
+        } else {
+            EffectDB::deleteById($_POST['delete_idEffect']);
+        }
+        
+        
+    }
 }
 
-if (isset($_POST['delete_cause']) && isset($_POST['delete_idCause'])) {
-    CauseDB::deleteCauseById($_POST['delete_idCause']);
+function errorHandlingDeleteCause(){
+    
+    if (isset($_POST['delete_cause']) && isset($_POST['delete_idCause'])) {
+    
+       if(ClusterDB::getCauseClusterWhereIdCause($_POST['delete_idCause']) && CauseEffectDB::getCauseEffectbyCauseForDelete($_POST['delete_idCause'])){
+            
+            echo '<span style="color:red">Delete denied: Cause exists in Cluster and in Cause-Effect</span>';
+            
+        } else if(ClusterDB::getCauseClusterWhereIdCause($_POST['delete_idCause'])){
+           
+           echo '<span style="color:red">Delete denied: Cause exists in Cluster</span>';
+       
+       } else if (CauseEffectDB::getCauseEffectbyCauseForDelete($_POST['delete_idCause'])){
+           
+           echo '<span style="color:red">Delete denied: Cause exists in Cause - Effect</span>';
+       } else {
+           CauseDB::deleteCauseById($_POST['delete_idCause']);
+       } 
+        
+    }
+
 }
+
 
 
 ?>
